Finish the controller adding the execute_movements method and put it in the folder api/v1
